handle touch events

+++ src/pkg/objects/position.go
@@ -4,10 +4,10 @@ import "fmt"
 
 // Position represents the position of an object (X, Y)
 type Position struct {
-	X, Y int
+	X, Y float64
 }
 
 // String returns the string representation of the position.
 func (pos Position) String() string {
-	return fmt.Sprintf("(%d, %d)", pos.X, pos.Y)
+	return fmt.Sprintf("(%g, %g)", pos.X, pos.Y)
 }
